refactor out the different location objects read that direct from controller when setting a waypoint location
